add AbstractSelectStatementExtractor
